Search part separation
