+++ src/components/MainPage/MainBanner.js
